// packages/datadog-api-client-v1/models/LogsExclusion.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { LogsExclusionFilter } from './LogsExclusionFilter';
/**
* Represents the index exclusion filter object from configuration API.
*/
export declare class LogsExclusion {
    'filter'?: LogsExclusionFilter;
    /**
    * Whether or not the exclusion filter is active.
    */
    'isEnabled'?: boolean;
    /**
    * Name of the index exclusion filter.
    */
    'name': string;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): LogsExclusion;
    static serialize(data: LogsExclusion): {
        [key: string]: any;
    };
    constructor();
}